pipeline/requirements.txt
# Core dependencies
langchain>=0.1.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
duckduckgo-search>=3.9.0
trafilatura>=1.6.0
httpx>=0.25.0
requests>=2.31.0
robotexclusionrulesparser>=1.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dateutil>=2.8.0

# Hugging Face
transformers>=4.30.0
torch>=2.0.0

# CLI and utilities
typer>=0.9.0
loguru>=0.7.0
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
